requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   watchfiles
attrs==26.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==5.0.0
    # via chromadb
beautifulsoup4==4.14.3
    # via -r requirements.in
build==1.5.0
    # via chromadb
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.7
    # via requests
chromadb==1.5.8
    # via
    #   -r requirements.in
    #   langchain-chroma
click==8.3.3
    # via
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
durationpy==0.10
    # via kubernetes
filelock==3.29.0
    # via huggingface-hub
flatbuffers==25.12.19
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.4.0
    # via huggingface-hub
googleapis-common-protos==1.74.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==3.5.0
    # via
    #   playwright
    #   sqlalchemy
grpcio==1.80.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
html2text==2025.4.15
    # via -r requirements.in
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   huggingface-hub
    #   langgraph-sdk
    #   langsmith
    #   ollama
httpx-sse==0.4.3
    # via langchain-community
huggingface-hub==1.13.0
    # via tokenizers
idna==3.13
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via opentelemetry-api
importlib-resources==7.1.0
    # via chromadb
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
jsonschema==4.26.0
    # via chromadb
jsonschema-specifications==2025.9.1
    # via jsonschema
kubernetes==35.0.0
    # via chromadb
langchain==1.2.17
    # via -r requirements.in
langchain-chroma==1.1.0
    # via -r requirements.in
langchain-classic==1.0.5
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.3.2
    # via
    #   langchain
    #   langchain-chroma
    #   langchain-classic
    #   langchain-community
    #   langchain-ollama
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-ollama==1.1.0
    # via -r requirements.in
langchain-protocol==0.0.15
    # via langchain-core
langchain-text-splitters==1.1.2
    # via
    #   -r requirements.in
    #   langchain-classic
langgraph==1.1.10
    # via langchain
langgraph-checkpoint==4.0.3
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.13
    # via langgraph
langgraph-sdk==0.3.13
    # via langgraph
langsmith==0.8.0
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
markdown-it-py==4.0.0
    # via rich
marshmallow==3.26.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.1
    # via chromadb
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.4.4
    # via
    #   chromadb
    #   langchain-chroma
    #   langchain-community
    #   onnxruntime
oauthlib==3.3.1
    # via requests-oauthlib
ollama==0.6.2
    # via
    #   -r requirements.in
    #   langchain-ollama
onnxruntime==1.25.1
    # via chromadb
opentelemetry-api==1.41.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.41.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.41.1
    # via chromadb
opentelemetry-proto==1.41.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.41.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.62b1
    # via opentelemetry-sdk
orjson==3.11.8
    # via
    #   chromadb
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
overrides==7.7.0
    # via chromadb
packaging==26.2
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   onnxruntime
playwright==1.59.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.6
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pybase64==1.4.3
    # via chromadb
pydantic==2.13.3
    # via
    #   chromadb
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
    #   pydantic-settings
pydantic-core==2.46.3
    # via pydantic
pydantic-settings==2.14.0
    # via
    #   chromadb
    #   langchain-community
pyee==13.0.1
    # via playwright
pygments==2.20.0
    # via rich
pypdf==6.10.2
    # via -r requirements.in
pypika==0.51.1
    # via chromadb
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via kubernetes
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   -r requirements.in
    #   kubernetes
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via langsmith
rich==15.0.0
    # via
    #   chromadb
    #   typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.49
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.4
    # via
    #   chromadb
    #   langchain-community
    #   langchain-core
tokenizers==0.23.1
    # via chromadb
tqdm==4.67.3
    # via
    #   chromadb
    #   huggingface-hub
typer==0.25.1
    # via
    #   chromadb
    #   huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   chromadb
    #   grpcio
    #   huggingface-hub
    #   langchain-core
    #   langchain-protocol
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   pyee
    #   referencing
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   kubernetes
    #   requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.46.0
    # via chromadb
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websocket-client==1.9.0
    # via kubernetes
websockets==16.0
    # via uvicorn
xxhash==3.7.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zipp==3.23.1
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
